feat: :sparkles: create dir for MapGenerator

+++ map-generator/MapGenerator-Python/main.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from io import TextIOWrapper
 import os
 import os.path
 
@@ -7,7 +6,7 @@ from easygui import multenterbox
 
 import CLR_IMPORT
 import SETTINGS
-from System import UInt32, String
+from System import String
 from GameClass.MapGenerator import MapStruct
 from Classes.MapRenderer import MapRenderer
 from Classes.RandomCores.PerlinRandomCore import PerlinRandomCore
@@ -19,7 +18,7 @@ if path[-len(SETTINGS.file_suffix):] != SETTINGS.file_suffix:
     path += SETTINGS.file_suffix
 # 地图加载
 if not os.path.exists(path):
-    height, width = [UInt32(int(x))
+    height, width = [int(x)
                      for x in multenterbox(msg='Create new map', title=SETTINGS.title, fields=['Height', 'Width'])]
     mapStruct = MapStruct(height, width)
     MapStruct.ToFile(String(path), mapStruct)
